Changes For added Security

+++ shops/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.http import HttpResponseForbidden
 from .models import Shop
 from .forms import ShopForm
 from django.views import View
@@ -10,8 +11,19 @@ from django.contrib.auth.decorators import user_passes_test
 from django.contrib.auth import views as auth_views
 from django.contrib.auth.decorators import login_required
 from django.utils.decorators import method_decorator
+from django.contrib.auth.forms import UserCreationForm
 
 
+
+def signup(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('login')  # Redirect to the login page after successful signup
+    else:
+        form = UserCreationForm()
+    return render(request, 'signup.html', {'form': form})
 
 @login_required
 def admin_panel(request):
@@ -19,7 +31,11 @@ def admin_panel(request):
 
 @login_required
 def shop_list(request):
-    shops = Shop.objects.all()
+    if request.user.is_superuser:
+        # Admin user can view all shops
+        shops = Shop.objects.all()
+    else:
+        shops = Shop.objects.filter(user=request.user)
     return render(request, 'shop_list.html', {'shops': shops})
 
 @login_required
@@ -36,19 +52,21 @@ def shop_create(request):
     if request.method == 'POST':
         form = ShopForm(request.POST)
         if form.is_valid():
-            print("saving")
-            shop = form.save()
+            shop = form.save(commit=False)
+            shop.user = request.user  # Set the user as the owner of the shop
+            shop.save()
             return redirect('shop_list')  # Redirect to shop list page
     else:
         form = ShopForm()
     return render(request, 'shop_create.html', {'form': form})
 
 @login_required
-@user_passes_test(lambda u: u.is_superuser)
+@user_passes_test(lambda u: u.is_superuser or u.shop_set.exists())
 @csrf_protect
 def shop_update(request, shop_id):
     shop = get_object_or_404(Shop, pk=shop_id)
-
+    if shop.user != request.user and not request.user.is_superuser:
+        return HttpResponseForbidden("You don't have permission to update this shop.")
     if request.method == 'POST':
         form = ShopForm(request.POST, instance=shop)
         if form.is_valid():
@@ -56,15 +74,16 @@ def shop_update(request, shop_id):
             return redirect('shop_detail', shop_id=shop_id)
     else:
         form = ShopForm(instance=shop)
-
     return render(request, 'shop_update.html', {'form': form, 'shop_id': shop_id})
 
 
 @login_required
-@user_passes_test(lambda u: u.is_superuser)
+@user_passes_test(lambda u: u.is_superuser or u.shop_set.exists())
 @csrf_protect
 def shop_delete(request, shop_id):
     shop = get_object_or_404(Shop, pk=shop_id)
+    if shop.user != request.user and not request.user.is_superuser:
+        return HttpResponseForbidden("You don't have permission to delete this shop.")
     if request.method == 'POST':
         shop.delete()
         return redirect('shop_list')
